Bladeテンプレートに含まれている <script> 部分を外部のjsファイルへ移した
一つ前のコミットの不具合は解消した。.envの右記設定に変更した。SESSION_DRIVER=file
